Make a separate module for linalg implementations
